seg_template/requirements.txt
torch
torchvision
torchaudio
opencv-python==4.8.1.78
tqdm==4.66.1
fastapi==0.103.2
uvicorn==0.23.2
streamlit==1.28.2
pandas==2.1.2
pillow>=9.0.0
plotly>=5.0.0
python-multipart>=0.0.6
segmentation-models-pytorch>=0.3.0
albumentations>=1.3.0
scikit-image>=0.19.0
matplotlib>=3.5.0
seaborn>=0.11.0
